Adding support for databinding.
Now actually getting databinding working. This requires using the ngOnChanges lifecycle hook to render when the input from the user changes. This also means I had to update the way I'm taking in the content for the label, because otherwise I wouldn't be aware of it and it wouldn't be updated. I'd like to come back and see if there is a better way to watch for content changes in the ng-content itself. This implementaiton really is relying on React's render function to be really smart, which from doing some intial perf testing it is.